Created AutoOpMode
NOTES:
NEW: ProductionOpMode. Filled the program with the rough idea of the program based of the template AutoOpMode already in the codebase.
REFACTORED: Robot.java -> Drivebase.java. All related variable names changed to reflect this change. I realized that it would be easier to not have an interface class to run autonomous, and that it would be easier to simply throw each mechanism into their own personal class. I'm keeping the old code for the robot class, but now it's more of doing the drivebase for TeleOp now.