Created Skeleton Mechanism Classes
NOTES:
NEW: Base Class. Created to abstract away all the nonsense for scheduling and actions from SolversLib. Since we are using the regular OpMode and not LinearOpmode, the CommandOpmode provided by the library is not available to use, and that we need to create and program all the boilerplate for ourselves.
NEW: Mechanism Classes (Camera, Flywheel, Intake). Created skeleton mechanism classes with planned methods to be programmed in future commits. Drivebase class might be rendered useless because of the useage of PedroPathing.
NEW: MecahnismsConstants. Created a class to store all of the gains for fusion algorithms for the future.